// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Programacion/Combinaciones.ts
export interface Combinaciones {
  clave: string;
  presentacion: string;
  liner1: string;
  medium1: string;
  liner2: string;
  medium2: string;
  liner3: string;
  medium3: string;
  liner4: string;
  primeraFlauta: string;
  segundaFlauta: string;
  terceraFlauta: string;
  pesoM2: number;
  descripcion: string;
  estatus: boolean;
  claveIngles: string;
  mullen: number;
}
export class Combinaciones {
  clave: string;
  presentacion: string;
  liner1: string;
  medium1: string;
  liner2: string;
  medium2: string;
  liner3: string;
  medium3: string;
  liner4: string;
  primeraFlauta: string;
  segundaFlauta: string;
  terceraFlauta: string;
  pesoM2: number;
  descripcion: string;
  estatus: boolean;
  claveIngles: string;
  mullen: number;
}